// src/types/types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  Date: { input: any; output: any; }
  DateTime: { input: any; output: any; }
  JSON: { input: any; output: any; }
  Long: { input: any; output: any; }
  Upload: { input: any; output: any; }
};
// ...
export enum Enum_Fiadocument_Type {
  Info = 'info',
  InvestigationNoFurtherAction = 'investigation_no_further_action',
  InvestigationPenaltyGridDrop = 'investigation_penalty_grid_drop',
  InvestigationPenaltyTime = 'investigation_penalty_time',
  InvestigationRightOfReviewPenaltyGridDrop = 'investigation_right_of_review_penalty_grid_drop',
  InvestigationRightOfReviewPenaltyNoFurtherAction = 'investigation_right_of_review_penalty_no_further_action',
  InvestigationRightOfReviewPenaltyTime = 'investigation_right_of_review_penalty_time'
}

export enum Enum_Schedulerscheduler_Type {
  Archive = 'archive',
  Publish = 'publish'
}

export enum Enum_Session_Type {
  Practice = 'practice',
  Qualifying = 'qualifying',
  Race = 'race',
  Session = 'session'
}

export enum Enum_Userlogin_Type {
  Driver = 'driver',
  Steward = 'steward'
}

export type FiaDocument = {
  __typename?: 'FiaDocument';
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  document?: Maybe<UploadFileRelationResponseCollection>;
  grand_prix?: Maybe<GrandPrixEntityResponse>;
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  session?: Maybe<SessionEntityResponse>;
  title?: Maybe<Scalars['String']['output']>;
  type?: Maybe<Enum_Fiadocument_Type>;
  uid?: Maybe<Scalars['String']['output']>;
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
  upload_date?: Maybe<Scalars['DateTime']['output']>;
};
// ...
export type FiaDocumentRelationResponseCollection = {
  __typename?: 'FiaDocumentRelationResponseCollection';
  data: Array<FiaDocumentEntity>;
};
// ...
export type GrandPrixEntity = {
  __typename?: 'GrandPrixEntity';
  attributes?: Maybe<GrandPrix>;
  id?: Maybe<Scalars['ID']['output']>;
};

export type GrandPrixEntityResponse = {
  __typename?: 'GrandPrixEntityResponse';
  data?: Maybe<GrandPrixEntity>;
};
// ...
export type Session = {
  __typename?: 'Session';
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  endTime?: Maybe<Scalars['DateTime']['output']>;
  fia_documents?: Maybe<FiaDocumentRelationResponseCollection>;
  grands_prix?: Maybe<GrandPrixEntityResponse>;
  name?: Maybe<Scalars['String']['output']>;
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  startTime?: Maybe<Scalars['DateTime']['output']>;
  type?: Maybe<Enum_Session_Type>;
  uid?: Maybe<Scalars['String']['output']>;
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
};


export type SessionFia_DocumentsArgs = {
  filters?: InputMaybe<FiaDocumentFiltersInput>;
  pagination?: InputMaybe<PaginationArg>;
  publicationState?: InputMaybe<PublicationState>;
  sort?: InputMaybe<Array<InputMaybe<Scalars['String']['input']>>>;
};

export type SessionEntity = {
  __typename?: 'SessionEntity';
  attributes?: Maybe<Session>;
  id?: Maybe<Scalars['ID']['output']>;
};

export type SessionEntityResponse = {
  __typename?: 'SessionEntityResponse';
  data?: Maybe<SessionEntity>;
};
// ...
export type UploadFileEntity = {
  __typename?: 'UploadFileEntity';
  attributes?: Maybe<UploadFile>;
  id?: Maybe<Scalars['ID']['output']>;
};
// ...
export type UploadFileRelationResponseCollection = {
  __typename?: 'UploadFileRelationResponseCollection';
  data: Array<UploadFileEntity>;
};